Fixed a prepared statement name clash.
When a find_by is executed with certain conditions, it is stored under a
generated prepared name.
If following that a find_by with the same conditions but adding a limit and
offset, the generated name is the same as the previous one, but the query
expects two additional arguments.
This commit makes the generated statement name depend on the limit as well.